[refactor] 게시글 오류 리펙토링
