+++ src/main/java/com/hunmin/domain/controller/ChatRoomController.java
@@ -38,12 +38,6 @@ public class ChatRoomController {
         return chatRoomService.findRoomByEmail(currentMemberEmail);
     }
 
-    //채팅방 안으로 입장
-    @GetMapping("/enter/{chatRoomId}")
-    @Operation(summary = "채팅방 입장", description = "채팅방으로 입장하는 API")
-    public String RoomEnter(@PathVariable String chatRoomId) {
-        return chatRoomId;
-    }
     //단일 채팅방 정보 조회
     @GetMapping("/{chatRoomId}")
     @Operation(summary = "단일 채팅방 정보 조회", description = "검색하고 싶은 채팅방을 조회하는 API")
