+++ src/main/java/com/hunmin/domain/dto/chat/ChatMessageListRequestDTO.java
@@ -1,5 +1,6 @@
 package com.hunmin.domain.dto.chat;
 
+import com.hunmin.domain.entity.MessageType;
 import lombok.Data;
 import lombok.NoArgsConstructor;
