+++ src/main/java/com/hunmin/domain/entity/MessageType.java
@@ -1,4 +1,4 @@
-package com.hunmin.domain.dto.chat;
+package com.hunmin.domain.entity;
 
 public enum MessageType {
         ENTER, QUIT, TALK
