+++ src/test/java/com/hunmin/domain/repository/ChatMessageRepositoryTest.java
@@ -2,7 +2,7 @@ package com.hunmin.domain.repository;
 
 import com.hunmin.domain.dto.chat.ChatMessageDTO;
 import com.hunmin.domain.dto.chat.ChatMessageListRequestDTO;
-import com.hunmin.domain.dto.chat.MessageType;
+import com.hunmin.domain.entity.MessageType;
 import com.hunmin.domain.entity.ChatMessage;
 import com.hunmin.domain.entity.ChatRoom;
 import com.hunmin.domain.entity.Member;
