+++ src/test/java/com/hunmin/domain/repository/ChatRoomRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.hunmin.domain.repository;
 
 import com.hunmin.domain.dto.chat.ChatRoomDTO;
-import com.hunmin.domain.dto.chat.MessageType;
+import com.hunmin.domain.entity.MessageType;
 import com.hunmin.domain.entity.*;
 import com.hunmin.domain.exception.ChatRoomException;
 import lombok.extern.log4j.Log4j2;
